+++ src/sdk/pynni/nni/nas/pytorch/search_space_zoo/__init__.py
@@ -1,4 +1,5 @@
 from .darts_cell import DartsCell
 from .enas_cell import ENASMicroLayer
-from .enas_cell import ENASReductionLayer
+# from .enas_cell import ENASReductionLayer
 from .enas_cell import ENASMacroLayer
+from .enas_cell import ENASMacroGeneralModel
