+++ src/sdk/pynni/nni/nas/pytorch/search_space_zoo/enas_cell.py
@@ -55,7 +55,7 @@ class Calibration(nn.Module):
 
 class ENASReductionLayer(nn.Module):
     """
-    builtin EnasReductionLayer. 
+    builtin EnasReductionLayer.
 
     Parameters
     ---
@@ -88,6 +88,8 @@ class ENASMicroLayer(nn.Module):
 
     Parameters
     ---
+    cells_list: list
+        the list contains all cells in model
     num_nodes: int
         the number of nodes contained in this layer
     in_channles_pp: int
@@ -99,8 +101,12 @@ class ENASMicroLayer(nn.Module):
     reduction: bool
         is previous layer a reduction layer
     """
-    def __init__(self, num_nodes, in_channels_pp, in_channels_p, out_channels, reduction):
+    def __init__(self, cells_list, num_nodes, in_channels_pp, in_channels_p, out_channels, reduction):
         super().__init__()
+        print(in_channels_pp, in_channels_p, out_channels, reduction)
+        if reduction:
+            cells_list.append(ENASReductionLayer(in_channels_pp, in_channels_p, out_channels))
+            in_channels_pp = in_channels_p = out_channels
         self.preproc0 = Calibration(in_channels_pp, out_channels)
         self.preproc1 = Calibration(in_channels_p, out_channels)
 
@@ -179,3 +185,70 @@ class ENASMacroLayer(mutables.MutableScope):
             if connection is not None:
                 out += connection
         return self.batch_norm(out)
+
+
+class ENASMacroGeneralModel(nn.Module):
+    """
+    The network is made up by stacking ENASMacroLayer. The Macro search space contains these layers.
+    Each layer chooses a operation from predefined ones and SkipConnect then form a network.
+
+    Parameters
+    ---
+    num_layers: int
+        The number of layers contained in the network.
+    out_filters: int
+        The number of each layer's output channels.
+    in_channel: int
+        The number of input's channels.
+    num_classes: int
+        The number of classes for classification.
+    dropout_rate: float
+        Dropout layer's dropout rate before the final dense layer.
+    """
+    def __init__(self, num_layers=12, out_filters=24, in_channels=3, num_classes=10,
+                 dropout_rate=0.0):
+        super().__init__()
+        self.num_layers = num_layers
+        self.num_classes = num_classes
+        self.out_filters = out_filters
+
+        self.stem = nn.Sequential(
+            nn.Conv2d(in_channels, out_filters, 3, 1, 1, bias=False),
+            nn.BatchNorm2d(out_filters)
+        )
+
+        pool_distance = self.num_layers // 3
+        self.pool_layers_idx = [pool_distance - 1, 2 * pool_distance - 1]
+        self.dropout_rate = dropout_rate
+        self.dropout = nn.Dropout(self.dropout_rate)
+
+        self.layers = nn.ModuleList()
+        self.pool_layers = nn.ModuleList()
+        labels = []
+        for layer_id in range(self.num_layers):
+            labels.append("layer_{}".format(layer_id))
+            if layer_id in self.pool_layers_idx:
+                self.pool_layers.append(FactorizedReduce(self.out_filters, self.out_filters))
+            self.layers.append(ENASMacroLayer(labels[-1], labels[:-1], self.out_filters, self.out_filters))
+
+        self.gap = nn.AdaptiveAvgPool2d(1)
+        self.dense = nn.Linear(self.out_filters, self.num_classes)
+
+    def forward(self, x):
+        bs = x.size(0)
+        cur = self.stem(x)
+
+        layers = [cur]
+
+        for layer_id in range(self.num_layers):
+            cur = self.layers[layer_id](layers)
+            layers.append(cur)
+            if layer_id in self.pool_layers_idx:
+                for i, layer in enumerate(layers):
+                    layers[i] = self.pool_layers[self.pool_layers_idx.index(layer_id)](layer)
+                cur = layers[-1]
+
+        cur = self.gap(cur).view(bs, -1)
+        cur = self.dropout(cur)
+        logits = self.dense(cur)
+        return logits
